Remove GIO dependency
This patch updates the code that saves and loads BDDs to use standard
C streams, rather than GIO streams.  This adds some complexity to our
test cases, since there isn't a portable interface for creating
memory-backed C streams — we have to use temporary files, instead.
But on the flip side, it means that the library works with older
versions of GLib — specifically, with the version that's available
under RHEL5 (2.13).