Use integer indices for BDD node IDs
Before, we used a pointer to the struct ipset_node instance as the node
ID for each nonterminal node in a BDD.  This patch changes things so
that we use an int index.  This has two benefits.  First, on 64-bit
platforms, it reduces the size of the ipset_node type, since we're
storing 32-bit indices instead of 64-bit pointers.  Second, it lets us
allocate the ipset_node instance in chunks, which should help with
memory locality.