Simplify BDD operator implementation
Before, we used some standard BDD operators to implement sets and maps.
For instance, to add an address to a set, we'd construct a BDD that
represented the address, and then use the BDD OR (∨) operator to add the
address's BDD to the set's BDD.  The result would be the new BDD for the
set.  (For maps, we'd use the if-then-else (ITE) operator, which
short-circuit ∨ is a special case of.)  We also used the standard
implementation strategy of memoizing the results of each of the BDD
operators.

This patch simplifies things by merging the two steps (construct a BDD
for the address; add it to the set) into a single operation, implemented
in the new `ipset_bdd_node_insert` function.

Doing this has a couple of benefits.  First, each individual insertion
is now slightly faster, since we only have to recurse through the
structure of the new IP address once.  (Before we had to do this once
while creating the address's BDD, and once while adding it to the set's
BDD.)

Second, and more importantly, it let's us get rid of the per-operator
caches.  This, in turn, will make it simpler to expire unused
nonterminal nodes once they're no longer needed.  And that, in turn,
will reduce the memory requirements of building up a large set.