Maintain a free list of reusable BDD nodes
The ipset_node_cache class now maintains a free list of nodes that are
no longer needed, and whose storage can be reused by later nodes.  We
maintain the free list using the indices into the chunks table; this
lets us reuse one of the fields in the node struct to store the “next”
pointer, rather than having to make the struct bigger.

This greatly reduces the storage requirements of the library, especially
when handling many large sets.  (Before, we would never reclaim or free
the node instances, which was a pretty massive memory leak.)