More literal translation of BDD APPLY operator
My previous implementation of the ipset_bdd_insert function didn't
handle the following input correctly:

  10.0.5.64
  10.0.5.66
  10.0.5.67

It would incorrectly produce the set containing 10.0.5.64/31 — i.e.,
10.0.5.65 would also be in the set, even though that wasn't in the
input.

The issue is that the set {10.0.5.64, 10.0.5.66} skips over one of the
bits in its BDD representation, since the two addresses are identical
except for the 0x0002 bit.  Before, when we tried to add 10.0.5.67, we
should have recreated the nonterminal node for that bit, since we'd have
to start checking it again in the new set.  But we didn't.

To get around this, I replaced the old implementation of
ipset_bdd_insert with one that more closely follows the standard
implementation of the APPLY operator in the BDD literature.  We still
don't directly create a BDD for the LHS of the operator (which is the
element that we're trying to add); instead, we have a “fake” BDD that we
can step through without having to physically create its BDD nodes.