--verbose option in all command-line tools
All of the command-line tools now take in a --verbose option.  If you provide
this option, the output of the command will be exactly the same as before.  If
you don't provide it, then a lot of the diagnostic/summary information is left
out.  (Basically, you'll only see error messages.)