Fix ipmap_set when setting element to 0
The BDD code did not correctly handle the case where an IP address was
previously set to a non-zero value, and then you tried to set that IP address to
0.  (Currently that particular use case is only available via ipmap_set, but it
would also crop up as part of a hypothetical ipset_remove function.)  The
problem was that we were using logical or (||) to construct the BDD for the new
set, and so the previous value would take precedence over the requested new 0
value.  We now use the if-then-else (ITE, ?:) operator to construct the new BDD.
For set addition, this is the same as logical or, and for maps, it implements
the correct behavior when overwriting old values.