Ignore order of input lines in ipsetbuild
The order of IP address additions and removes in ipsetbuild is ignored by
creating a cork array to hold all removals records while reading the input
file.  Additions are handled as they are read.  After the file has been read,
the cork array is traversed to handle all removal lines.  This ensures that we
get the same result set regardless of what order the lines appear in the input
files.

Includes idempotent tests cases in `tests/round-trip`.