Rename ipset → libcorkipset
`ipset` is pretty generic, and already taken.  Arch and Debian both
renamed this to libcorkipset, so that's what we'll do too!

I've cleaned up the header files a bit; the main one is now installed as
<libcork/ipset.h> instead of <libcorkipset/ipset.h>, which is what's
used in the current Arch and Debian packages.  I've kept a copy
available at the old location, with a deprecation warning telling you
about the new location.